Update
Внес правки, указанные ранее.
Теперь контроллеры для отправки/приема используют общий контроллер, таким образом убрал дублирование кода.
Сервер также теперь использует единый метод Send для отправки.
Клиенты теперь хранятся в одной коллекции, а не в трех.
Сообщения чата хранятся в листе, а не единой строкой.
Убрал постоянные подписки/отписки событий, теперь подписка происходит лишь единожды в конструкторе.
Переделал работу с событиями, теперь там используется enum с флагами.

+++ Server/Settings.cs
@@ -19,7 +19,7 @@
         public int Port { get; set; }
 
         [XmlElement("timeout")]
-        public int Timeout { get; set; }
+        public long Timeout { get; set; }
 
         [XmlElement("dbName")]
         public string DbName { get; set; }
